+++ np.Tidy/CHANGELOG.md
@@ -1,6 +1,9 @@
 # 🧹 Tidy Up Changelog
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/np.Tidy/README.md) for full details on the available commands and use from callbacks and templates.
 
+## [1.19.1] - 2026-04-14 @jgclark
+- fix to '/Remove empty lines' command (thanks, @grdn)
+
 ## [1.19.0] - 2026-03-26 @jgclark
 ### New
 - new **/Cancel incomplete tasks in a past year** command that, for a year of your choice, counts and logs incomplete tasks and checklists in past calendar notes (daily, weekly, monthly, quarterly and yearly), shows per-note-type and per-Teamspace counts with a strong warning, and then (if confirmed) bulk-cancels those items.